feat: update StockMovement

+++ springboot-app/src/main/java/com/lumos/lumosspring/execution/controller/dto/ItemsResponse.java
@@ -1,4 +1,4 @@
 package com.lumos.lumosspring.execution.controller.dto;
 
-public record ItemsResponse(long materialId, String materialName, float materialQuantity) {
+public record ItemsResponse(long materialId, String materialName, double materialQuantity) {
 }
